Create DFG config and reorganize DFG folder
