+++ thoth/app/dfg/config.py
@@ -1,0 +1,8 @@
+class DFGConfig:
+    """
+    DFG Config
+    """
+
+    FONT = "Helvetica, Arial, sans-serif"
+    FONTSIZE = "20"
+    CONSTANT_NODE_COLOR = "lightblue"
